restructure msg lib for public api and move binary subproject to dedicated directory
