+++ binchihua/src/common/opt.h
