+++ libchihua-msg/include/common/memory.h
@@ -1,0 +1,6 @@
+#ifndef _BLUE_CHIHUAHUA_LIB_COMMON_MEMORY
+#define _BLUE_CHIHUAHUA_LIB_COMMON_MEMORY
+
+#define _cleanup_(x) __attribute__((__cleanup__(x)))
+
+#endif
