+++ libchihua-msg/include/orchestrator/controller.h
@@ -1,9 +1,10 @@
-#ifndef _BLUE_CHIHUAHUA_ORCHESTRATOR_CONTROLLER
-#define _BLUE_CHIHUAHUA_ORCHESTRATOR_CONTROLLER
+#ifndef _BLUE_CHIHUAHUA_LIB_ORCHESTRATOR_CONTROLLER
+#define _BLUE_CHIHUAHUA_LIB_ORCHESTRATOR_CONTROLLER
 
-#include "../../common/dbus.h"
+#include "../common/memory.h"
 
 #include <stdbool.h>
+#include <systemd/sd-event.h>
 
 // TODO: include common/dbus and cleanup event_source in _new
