+++ libchihua-msg/src/common/memory.h
@@ -1,6 +1,8 @@
 #ifndef _BLUE_CHIHUAHUA_COMMON_MEMORY
 #define _BLUE_CHIHUAHUA_COMMON_MEMORY
 
+#include "../../include/common/memory.h"
+
 #include <stdlib.h>
 #include <unistd.h>
 
@@ -22,8 +24,6 @@ static inline void freep(void *p) {
 }
 
 #define malloc0(n) (calloc(1, (n) ?: 1))
-
-#define _cleanup_(x) __attribute__((__cleanup__(x)))
 
 #define _cleanup_free_ _cleanup_(freep)
 #define _cleanup_fd_ _cleanup_(closep)
